file/requirements.txt
numpy==1.24.3
opencv_contrib_python==4.9.0.80
opencv_python==4.10.0.82
Pillow==10.0.0
Pillow==10.3.0
protobuf==5.27.1
PyGetWindow==0.0.9
pytesseract==0.3.10
pywin32==306
